% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkBoolean}
\alias{checkBoolean}
\title{Helper function to check parameter vectors for positive signed integer values}
\usage{
checkBoolean(parameter, name)
}
\description{
Helper function to check parameter vectors for positive signed integer values
}
\keyword{internal}
