% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkParameterMatrix}
\alias{checkParameterMatrix}
\title{Helper function to check parameter matrices}
\usage{
checkParameterMatrix(parameter, name, no_of_rows, no_of_cols)
}
\description{
Helper function to check parameter matrices
}
\keyword{internal}
