% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkPositiveDouble}
\alias{checkPositiveDouble}
\title{Helper function to check positive double function parameter}
\usage{
checkPositiveDouble(parameter, name)
}
\description{
Helper function to check positive double function parameter
}
\keyword{internal}
