% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkPositiveSignedInteger}
\alias{checkPositiveSignedInteger}
\title{Helper function to check unsigned integer function parameter}
\usage{
checkPositiveSignedInteger(parameter, name, bit_size = 64)
}
\description{
Helper function to check unsigned integer function parameter
}
\keyword{internal}
