% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkPositiveSignedParameterVector}
\alias{checkPositiveSignedParameterVector}
\title{Helper function to check parameter vectors for positive signed integer values}
\usage{
checkPositiveSignedParameterVector(parameter, name, size)
}
\description{
Helper function to check parameter vectors for positive signed integer values
}
\keyword{internal}
