% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factor_model}
\alias{factor_model}
\title{Example factor model dataset}
\format{
A \code{SimulData} containing the following elements:
\describe{
\item{data}{If \code{starting_date} is provided, a \code{zoo} object, else,
a (\code{no_of_vars} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding the
simulated data.}
\item{factors}{If \code{starting_date} is provided, a \code{zoo} object, else a
(\code{no_of_factors} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding  the
simulated latent factors.}
\item{trans_var_coeff}{Numeric (\code{no_of_factors} \eqn{\times}{x}
(\code{no_of_factors} * \code{factor_lag_order})) factor VAR coefficient matrix.}
\item{loading_matrix}{Numeric factor loading matrix.}
\item{meas_error}{If \code{starting_date} is provided, a \code{zoo} object, else
a (\code{no_of_vars} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding the
fundamental measurement errors.}
\item{meas_error_var_cov}{Numeric measurement error
variance-covariance matrix.}
\item{trans_error_var_cov}{Numeric transition error
variance-covariance matrix.}
\item{frequency}{Integer vector of variable frequencies.}
\item{delay}{Integer vector of variable delays, measured as the number of
months since the latest available observation.}
}
}
\source{
Generated via
seed <- 02102025
set.seed(seed)
no_of_obs <- 300
no_of_vars <- 50
no_of_factors <- 3
trans_error_var_cov <- diag(1, no_of_factors)
loading_matrix <- matrix(rnorm(no_of_vars * no_of_factors), no_of_vars, no_of_factors)
meas_error_mean <- rep(0, no_of_vars)
meas_error_var_cov <- diag(1, no_of_vars)
trans_var_coeff <- cbind(diag(0.5, no_of_factors), -diag(0.25, no_of_factors))
factor_lag_order <- 2
delay <- c(2, 1, floor(rexp(no_of_vars - 2, 1)))
quarterfy <- FALSE
quarterly_variable_ratio <- 0
corr <- TRUE
beta_param <- 2
burn_in <- 999
starting_date <- "1970-01-01"
rescale <- FALSE
check_stationarity <- TRUE
stationarity_check_threshold <- 1e-10
factor_model <- simFM(no_of_obs = no_of_obs, no_of_vars = no_of_vars,
no_of_factors = no_of_factors, loading_matrix = loading_matrix,
meas_error_mean = meas_error_mean, meas_error_var_cov = meas_error_var_cov,
trans_error_var_cov = trans_error_var_cov, trans_var_coeff = trans_var_coeff,
factor_lag_order = factor_lag_order, delay = delay, quarterfy = quarterfy,
quarterly_variable_ratio  = quarterly_variable_ratio, corr = corr,
beta_param = beta_param, seed = seed, burn_in = burn_in, starting_date = starting_date,
rescale = rescale, check_stationarity = check_stationarity,
stationarity_check_threshold = stationarity_check_threshold)
}
\usage{
factor_model
}
\description{
This is a simulated factor model dataset for demonstration and testing
with \code{TwoStepSDFM}.
}
\keyword{datasets}
