% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecaster.R
\name{forecastWrapper}
\alias{forecastWrapper}
\title{Internal forecasting wrapper function}
\usage{
forecastWrapper(
  target_variables,
  quarterly_predictors,
  factors,
  target_variable_delay,
  quarterly_delay,
  lag_estim_criterion,
  max_fcast_horizon,
  max_ar_lag_order,
  max_predictor_lag_order,
  jitter
)
}
\description{
This function is for internal use only and may change in future releases
without notice. Users should use \code{nowcast()} instead for a stable and
supported interface.
Helper function to check parameter vectors for positive signed integer values
}
\keyword{internal}
