% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeData.R
\name{imputeMonthlyData}
\alias{imputeMonthlyData}
\title{Impute monthly data.}
\usage{
imputeMonthlyData(
  data,
  delay,
  no_of_factors,
  max_factor_lag_order = 10,
  lag_estim_criterion = "BIC",
  decorr_errors = TRUE,
  comp_null = 1e-15,
  parallel = FALSE,
  fcast_horizon = 0,
  jitter = 1e-08
)
}
\arguments{
\item{data}{Numeric (no_of_vars \eqn{\times}{x} no_of_obs) matrix of data or
zoo/xts object sampled at the same frequency.}

\item{delay}{Integer vector of variable delays.}

\item{no_of_factors}{Integer number of factors.}

\item{max_factor_lag_order}{Integer maximum order of the VAR process in the
transition equation.}

\item{lag_estim_criterion}{Information criterion used for the estimation of
the factor VAR order (\code{"BIC"} (default), \code{"AIC"}, \code{"HIC"}).}

\item{decorr_errors}{Logical, whether or not the errors should be
decorrelated.}

\item{comp_null}{Numeric computational zero.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel
matrix operations.}

\item{fcast_horizon}{Integer number of additional Filter predictions into the
future.}

\item{jitter}{Numerical jitter for stability of internal solver algorithms.
The jitter is added to the diagonal entries of the variance covariance matrix
of the measurement errors.}
}
\value{
A named list containing the following objects:
\describe{
\item{imputed_data}{Imputed dataset with class inherited from \code{data}.}
\item{predicted_data}{In-sample model predictions with class inherited from
\code{data}.}
\item{dfm_fit}{\code{SDFMFit} object of the DFM fit.}
}
}
\description{
Impute missing data in monthly datasets using matrix imputation via a DFM
fit,
}
\details{
The function performs a two-step estimation procedure for dense dynamic
factor models as described in \insertRef{Giannone2008Nowcasting}{TwoStepSDFM}
and \insertRef{Doz2011Two_step}{TwoStepSDFM}. For more details, see
\code{\link{twoStepDenseDFM}}. The resulting estimates of the loading matrix
and factors are then used to predict any missing observations outside the
ragged edges.
}
\examples{
data(factor_model)
no_of_vars <- dim(factor_model$data)[2]
no_of_factors <- dim(factor_model$factors)[2]
factor_model$data[1:3, ] <- NA
factor_model$data
imputed_data <- imputeMonthlyData(data = factor_model$data, delay = factor_model$delay,
                                  no_of_factors = no_of_factors)
imputed_data$imputed_data

}
\references{
\insertRef{Giannone2008Nowcasting}{TwoStepSDFM}

\insertRef{Doz2011Two_step}{TwoStepSDFM}
}
\author{
Domenic Franjic
}
