% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{makeRaggedEdges}
\alias{makeRaggedEdges}
\title{Helper function to check positive double function parameter}
\usage{
makeRaggedEdges(data, delay)
}
\description{
Helper function to check positive double function parameter
}
\keyword{internal}
