% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{nowcastSpecificationHelper}
\alias{nowcastSpecificationHelper}
\title{Helper function to wrap the nowcasting routine}
\usage{
nowcastSpecificationHelper(
  cv_repetitions,
  no_of_factors,
  no_of_variables,
  no_of_observations,
  no_of_mtly_variables,
  lasso_penalty_type,
  data,
  variable_of_interest,
  fcast_horizon,
  delay,
  candidates,
  frequency,
  max_factor_lag_order,
  decorr_errors,
  lag_estim_criterion,
  max_iterations,
  comp_null,
  spca_conv_crit,
  max_ar_lag_order,
  max_predictor_lag_order,
  jitter,
  svd_method,
  weights
)
}
\description{
Helper function to wrap the nowcasting routine
}
\keyword{internal}
