% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepDenseDFM.R
\name{plot.DFMFit}
\alias{plot.DFMFit}
\title{Generic plotting function for DFMFit S3 objects}
\usage{
\method{plot}{DFMFit}(x, ...)
}
\arguments{
\item{x}{\code{DFMFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list of plot objects:
\describe{
\item{\verb{Factor Time Series Plots}}{\code{patchwork}/\code{ggplot} object showing the estimated factors over time with 95\% confidence bands of the Kalman Filter and Smoother.}
\item{\verb{Loading Matrix Heatmap}}{\code{ggplot} object showing a heatmap of the estimated factor loadings.}
\item{\verb{Meas. Error Var.-Cov. Matrix Heatmap}}{\code{ggplot} object showing a heatmap of the measurement error variance–covariance matrix.}
}
}
\description{
Create diagnostic plots for a \code{DFMFit} object, including factor time series plots,
the loading matrix heatmap and the measurement error variance–covariance matrix heatmap.
}
\examples{
\dontrun{
set.seed(02102025)
no_of_observations <- 100
no_of_variables <- 20
no_of_factors <- 3
trans_error_var_cov <- diag(1, no_of_factors)
loading_matrix <- matrix(round(rnorm(no_of_variables * no_of_factors)), no_of_variables, no_of_factors)
meas_error_mean <- rep(0, no_of_variables)
meas_error_var_cov <- diag(1, no_of_variables)
trans_var_coeff <- cbind(diag(0.5, no_of_factors), -diag(0.25, no_of_factors))
factor_lag_order <- 2
FM <- simFM(no_of_observations = no_of_observations, no_of_variables = no_of_variables, 
            no_of_factors = no_of_factors, loading_matrix = loading_matrix, 
            meas_error_mean = meas_error_mean, meas_error_var_cov = meas_error_var_cov,
            trans_error_var_cov = trans_error_var_cov, trans_var_coeff = trans_var_coeff, 
            factor_lag_order = factor_lag_order)

# Estimate a dense 2-factor model
fit <- twoStepDenseDFM(
  data = FM$data,
  delay = FM$delay,
  no_of_factors = no_of_factors
)

# Inspect the results
plots <- plot(fit)
plots$`Factor Time Series Plots`
plots$`Loading Matrix Heatmap`
plots$`Meas. Error Var.-Cov. Matrix Heatmap`
}

}
