% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KalmanFilterSmoother.R
\name{plot.KFSFit}
\alias{plot.KFSFit}
\title{Generic plotting function for KFSFit S3 objects}
\usage{
\method{plot}{KFSFit}(x, axis_text_size = 20, legend_title_text_size = 20, ...)
}
\arguments{
\item{x}{\code{KFSFit} object.}

\item{axis_text_size}{Numeric size of x- and y-axis labels. Prased to ggplot2
\code{theme(..., text = element_text(size = axis_text_size))}.}

\item{legend_title_text_size}{Numeric size of x- and y-axis labels. Prased to
ggplot2
\code{theme(..., legend.title = element_text(size = legend_title_text_size))}.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list of \code{patchwork}/\code{ggplot} objects:
\describe{
\item{\verb{Factor Time Series Plots}}{\code{patchwork}/\code{ggplot} object graphing the
estimated factors over time with 95\% confidence bands based on the smoother
uncertainty of the Kalman Filter and Smoother.}
}
}
\description{
Create diagnostic plots for a \code{KFSFit} object.
}
\author{
Domenic Franjic
}
