% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateFactors.R
\name{plot.NoOfFactorsFit}
\alias{plot.NoOfFactorsFit}
\title{Generic plotting function for NoOfFactorsFit S3 objects}
\usage{
\method{plot}{NoOfFactorsFit}(x, axis_text_size = 20, legend_title_text_size = 20, ...)
}
\arguments{
\item{x}{\code{NoOfFactorsFit} object.}

\item{axis_text_size}{Numeric size of x- and y-axis labels. Prased to ggplot2
\code{theme(..., text = element_text(size = axis_text_size))}.}

\item{legend_title_text_size}{Numeric size of x- and y-axis labels. Prased to
ggplot2
\code{theme(..., legend.title = element_text(size = legend_title_text_size))}.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list of plot objects:
\describe{
\item{\verb{Eigen Value Plot}}{\code{ggplot} object showing a bar plot of the
eigenvalues of the complex data Gram matrix.}
}
}
\description{
Create diagnostic plots for an \code{NoOfFactorsFit} object,
}
\author{
Domenic Franjic
}
