% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{plot.SDFMcrossVal}
\alias{plot.SDFMcrossVal}
\title{Generic plotting function for SDFMcrossVal S3 objects}
\usage{
\method{plot}{SDFMcrossVal}(x, axis_text_size = 20, legend_title_text_size = 20, ...)
}
\arguments{
\item{x}{\code{SDFMcrossVal} object.}

\item{axis_text_size}{Numeric size of x- and y-axis labels. Prased to ggplot2
\code{theme(..., text = element_text(size = axis_text_size))}.}

\item{legend_title_text_size}{Numeric size of x- and y-axis labels. Prased to
ggplot2
\code{theme(..., legend.title = element_text(size = legend_title_text_size))}.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list of \code{ggplot} objects:
\describe{
\item{\verb{CV Results}}{\code{ggplot} object of the cross-validation error against
the log Ridge penalty. The overall sparsity level of the loading matrix
induced by the lasso penalty is indicated by point shapes and colours.}
\item{\verb{BIC Results}}{\code{ggplot} object of the BIC against the log Ridge
penalty. The overall sparsity level of the loading matrix induced by the
lasso penalty is indicated by point shapes and colours.}
}
}
\description{
Generic plotting function for SDFMcrossVal S3 objects
}
\author{
Domenic Franjic
}
