% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{plot.SDFMFit}
\alias{plot.SDFMFit}
\title{Generic plotting function for SDFMFit S3 objects}
\usage{
\method{plot}{SDFMFit}(x, axis_text_size = 20, legend_title_text_size = 20, ...)
}
\arguments{
\item{x}{\code{SDFMFit} object.}

\item{axis_text_size}{Numeric size of x- and y-axis labels. Passed to ggplot2
\code{theme(..., text = element_text(size = axis_text_size))}.}

\item{legend_title_text_size}{Numeric size of x- and y-axis labels. Passed to
ggplot2
\code{theme(..., legend.title = element_text(size = legend_title_text_size))}.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list of plot objects:
\describe{
\item{\verb{Factor Time Series Plots}}{\code{patchwork}/\code{ggplot} object graphing the
estimated factors over time with 95\% confidence bands based on the smoother
uncertainty of the Kalman Filter and Smoother.}
\item{\verb{Loading Matrix Heatmap}}{\code{ggplot} object showing a heatmap of the
estimated factor loadings. Zeros are highlighted in black.}
\item{\verb{Meas. Error Var.-Cov. Matrix Heatmap}}{\code{ggplot} object showing a
heatmap of the measurement error variance-covariance matrix.}
\item{\verb{Eigenvalue Plot}}{\code{ggplot} object showing a bar plot of the
eigenvalues of the measurement error variance–covariance matrix.}
}
}
\description{
Generic plotting function for SDFMFit S3 objects
}
\author{
Domenic Franjic
}
