% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nowcast.R
\name{plot.SDFMnowcast}
\alias{plot.SDFMnowcast}
\title{Generic plotting function for SDFMnowcast S3 objects}
\usage{
\method{plot}{SDFMnowcast}(x, axis_text_size = 20, ...)
}
\arguments{
\item{x}{\code{SDFMnowcast} object.}

\item{axis_text_size}{Numeric size of x- and y-axis labels. Prased to ggplot2
\code{theme(..., text = element_text(size = axis_text_size))}.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
A named list storing of \code{ggplot} objects:
\describe{
\item{\verb{Single Pred. Fcast Density Plots x}}{\code{patchwork} / \code{ggplot} objects
graphing the distribution of forecasts generated by the predictors for
each prediction (backcasts, nowcasts, forecasts) for each target,
respectively. Altogether, there will be as many such objects as there are
targets, with \code{x} replaced by the column name of the target.}
}
}
\description{
Generic plotting function for SDFMnowcast S3 objects
}
\author{
Domenic Franjic
}
