% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFM.r
\name{plot.simFM}
\alias{plot.simFM}
\title{Generic plotting function for simFM S3 objects}
\usage{
\method{plot}{simFM}(x, ...)
}
\arguments{
\item{x}{\code{simFM} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic plotting function for simFM S3 objects
}
