% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotterHelper.R
\name{plotFactorEstimates}
\alias{plotFactorEstimates}
\title{Helper for plotting factor time series}
\usage{
plotFactorEstimates(
  factors,
  smoothed_state_variance,
  no_of_factors,
  axis_text_size
)
}
\description{
Helper for plotting factor time series
}
\keyword{internal}
