% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotterHelper.R
\name{plotInformationCrit}
\alias{plotInformationCrit}
\title{Helper for plotting ICs for selecting the number of factors}
\usage{
plotInformationCrit(information_crit, no_of_factors, axis_text_size)
}
\description{
Helper for plotting ICs for selecting the number of factors
}
\keyword{internal}
