% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotterHelper.R
\name{plotLoadingHeatMap}
\alias{plotLoadingHeatMap}
\title{Helper for plotting loading matrix heat maps}
\usage{
plotLoadingHeatMap(
  loading_matrix_estim,
  series_names,
  no_of_factors,
  axis_text_size,
  legend_title_text_size
)
}
\description{
Helper for plotting loading matrix heat maps
}
\keyword{internal}
