% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotterHelper.R
\name{plotMeasVarCovEigenvalues}
\alias{plotMeasVarCovEigenvalues}
\title{Helper for plotting measurement error var.cov. eigenvalues}
\usage{
plotMeasVarCovEigenvalues(
  eigen_values,
  no_of_factors,
  axis_text_size,
  legend_title_text_size
)
}
\description{
Helper for plotting measurement error var.cov. eigenvalues
}
\keyword{internal}
