% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KalmanFilterSmoother.R
\name{print.KFSFit}
\alias{print.KFSFit}
\title{Generic printing function for KFSFit S3 objects}
\usage{
\method{print}{KFSFit}(x, ...)
}
\arguments{
\item{x}{\code{KFSFit} object.}

\item{...}{Additional parameters.}
}
\value{
No return value, called for side effects.
}
\description{
Print a compact summary of a \code{KFSFit} object.
}
\author{
Domenic Franjic
}
