% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateFactors.R
\name{print.NoOfFactorsFit}
\alias{print.NoOfFactorsFit}
\title{Generic printing function for NoOfFactorsFit S3 objects}
\usage{
\method{print}{NoOfFactorsFit}(x, ...)
}
\arguments{
\item{x}{\code{NoOfFactorsFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
No return value; Prints a summary to the console.
}
\description{
Print a compact summary of an \code{NoOfFactorsFit} object.
}
\author{
Domenic Franjic
}
