% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{print.SDFMcrossVal}
\alias{print.SDFMcrossVal}
\title{Generic print function for SDFMcrossVal S3 objects}
\usage{
\method{print}{SDFMcrossVal}(x, ...)
}
\arguments{
\item{x}{\code{SDFMcrossVal} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
No return value; Prints a summary to the console.
}
\description{
Generic print function for SDFMcrossVal S3 objects
}
\author{
Domenic Franjic
}
