% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{print.SDFMFit}
\alias{print.SDFMFit}
\title{Generic printing function for SDFMFit S3 objects}
\usage{
\method{print}{SDFMFit}(x, ...)
}
\arguments{
\item{x}{\code{SDFMFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
No return value; Prints a summary to the console.
}
\description{
Print a compact summary of an \code{SDFMFit} object.
}
\author{
Domenic Franjic
}
