% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFM.r
\name{print.simFM}
\alias{print.simFM}
\title{Generic printing function for simFM S3 objects}
\usage{
\method{print}{simFM}(x, ...)
}
\arguments{
\item{x}{\code{simFM} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic printing function for simFM S3 objects
}
