% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFM.r
\name{print.SimulData}
\alias{print.SimulData}
\title{Generic printing function for SimulData S3 objects}
\usage{
\method{print}{SimulData}(x, ...)
}
\arguments{
\item{x}{\code{SimulData} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
No return value; Prints a summary to the console.
}
\description{
Print a compact summary of an \code{SimulData} object.
}
\author{
Domenic Franjic
}
