% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparcePCA.R
\name{print.SPCAFit}
\alias{print.SPCAFit}
\title{Generic printing function for SPCAFit S3 objects}
\usage{
\method{print}{SPCAFit}(x, ...)
}
\arguments{
\item{x}{\code{SPCAFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\value{
No return value; Prints a summary to the console.
}
\description{
Print a compact summary of an \code{SPCAFit} object.
}
\author{
Domenic Franjic
}
