% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepDenseDFM.R
\name{twoStepDenseDFM}
\alias{twoStepDenseDFM}
\title{Two Step Dense Dynamic Factor Model Estimator.}
\usage{
twoStepDenseDFM(
  data,
  delay,
  no_of_factors,
  max_factor_lag_order = 10,
  lag_estim_criterion = "BIC",
  decorr_errors = TRUE,
  comp_null = 1e-15,
  parallel = FALSE,
  fcast_horizon = 0,
  jitter = 1e-08
)
}
\arguments{
\item{data}{Numeric (no_of_vars \eqn{\times}{x} no_of_obs) matrix of data or
zoo/xts object sampled at the same frequency.}

\item{delay}{Integer vector of variable delays.}

\item{no_of_factors}{Integer number of factors.}

\item{max_factor_lag_order}{Integer maximum order of the VAR process in the
transition equation.}

\item{lag_estim_criterion}{Information criterion used for the estimation of
the factor VAR order (\code{"BIC"} (default), \code{"AIC"}, \code{"HIC"}).}

\item{decorr_errors}{Logical, whether or not the errors should be
decorrelated.}

\item{comp_null}{Numeric computational zero.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel
matrix operations.}

\item{fcast_horizon}{Integer number of additional Filter predictions into the
future.}

\item{jitter}{Numerical jitter for stability of internal solver algorithms.
The jitter is added to the diagonal entries of the variance covariance matrix
of the measurement errors.}
}
\value{
An object of class SDFMFit with main components:
\describe{
\item{data}{Original data object.}
\item{loading_matrix_estim}{Numeric matrix of estimated factor loadings.}
\item{smoothed_factors}{Object containing the SPCA factor estimates. The
object inherits its class from data: If data is provided as \code{zoo},
\code{factor_estim} will be a \code{zoo} object. If data is provided as matrix,
\code{factor_estim} will be a (\code{no_of_factors}\eqn{\times}{x}\code{no_of_obs}
matrix.}
\item{smoothed_state_variance}{(\code{no_of_factors}\eqn{\times}{x}(
\code{no_of_factors} * \code{no_of_obs})) matrix, where each (\code{no_of_factors}
\eqn{\times}{x}\code{no_of_factors}) block represents the smoother uncertainty
at time point\eqn{t}{t}.}
\item{factor_var_lag_order}{Integer order of the VAR process in the state
equation.}
\item{error_var_cov_cholesky_factor}{Numeric lower-triangular Cholesky
factor of the estimated measurement error variance–covariance matrix.}
\item{llt_success_code}{Integer indicating the status of the Cholesky
factorization: \code{0} = LLT succeeded, \code{-1} = LLT failed but LDLT succeeded,
\code{-2} = both failed and errors are treated as uncorrelated.}
}
}
\description{
Estimate a dense dynamic factor model with measurement equation
\deqn{
  \bm{x}_t = \bm{\Lambda} \bm{f}_{t} + \bm{\xi}_t,\quad \bm{\xi}_t \sim \mathcal{N}(\bm{0}, \bm{\Sigma}_{\xi}),
}
and transition equation
\deqn{
  \bm{f}_t = \sum_{p=0}^P\bm{\Phi}_p \bm{f}_{t-p} + \bm{\epsilon}_t,\quad \bm{\epsilon}_t \sim \mathcal{N}(\bm{0}, \bm{\Sigma}_{f}).
}
using principal components  analysis and the Kalman Filter and
Smoother according to \insertRef{Giannone2008Nowcasting}{TwoStepSDFM} and
\insertRef{Doz2011Two_step}{TwoStepSDFM}.
}
\details{
The function performs a two-step estimation procedure for dense dynamic
factor models as described in \insertRef{Giannone2008Nowcasting}{TwoStepSDFM}
and \insertRef{Doz2011Two_step}{TwoStepSDFM}. In the first step, the factor
loading matrix is estimated using PCA. In the second step the latent factors
are estimated using the univariate representation of the Kalman Filter and
Smoother \insertCite{koopman2000fast}{TwoStepSDFM}.

With respect to the univariate representation of the Kalman filter and
smoother, \code{decorr_errors} indicates whether the data should be decorrelated
internally prior to filtering and smoothing. \code{jitter} is added to the
diagonal elements of the measurement variance–covariance matrix. For more
details, see \code{\link{kalmanFilterSmoother}}.
}
\examples{
data(factor_model)
no_of_vars <- dim(factor_model$data)[2]
no_of_factors <- dim(factor_model$factors)[2]
dfm_fit <- twoStepDenseDFM(data = factor_model$data, delay = factor_model$delay, 
                           no_of_factors = no_of_factors)
print(dfm_fit)
dfm_plots <- plot(dfm_fit)
dfm_plots$`Factor Time Series Plots`
dfm_plots$`Loading Matrix Heatmap`
dfm_plots$`Meas. Error Var.-Cov. Matrix Heatmap`
dfm_plots$`Meas. Error Var.-Cov. Eigenvalue Plot`

}
\references{
\insertRef{koopman2000fast}{TwoStepSDFM}

\insertRef{Giannone2008Nowcasting}{TwoStepSDFM}

\insertRef{eigenweb}{TwoStepSDFM}

\insertRef{Doz2011Two_step}{TwoStepSDFM}
}
\seealso{
\code{\link{sparsePCA}}: Routine for fitting estimating a sparse factor
loading matrix.

\code{\link{kalmanFilterSmoother}}: Routine for filtering and smoothing
latent factors.

\code{\link{twoStepSDFM}}: Two-step estimation routine for a sparse dynamic
factor model.
}
\author{
Domenic Franjic
}
