% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{crossVal}
\alias{crossVal}
\title{Cross-validate the hyper-parameter of the mixed-frequency SDFM}
\usage{
crossVal(
  data,
  variable_of_interest,
  fcast_horizon,
  delay,
  frequency,
  no_of_factors,
  seed,
  min_ridge_penalty,
  max_ridge_penalty,
  cv_repititions,
  cv_size,
  lasso_penalty_type,
  min_max_penalty,
  max_factor_lag_order = 10,
  decorr_errors = TRUE,
  lag_estim_criterion = "BIC",
  ridge_penalty = 1e-06,
  lasso_penalty = NULL,
  max_iterations = 1000,
  max_no_steps = NULL,
  comp_null = 1e-15,
  check_rank = FALSE,
  conv_crit = 1e-04,
  conv_threshold = 1e-04,
  parallel = FALSE,
  max_ar_lag_order = 5,
  max_predictor_lag_order = 5
)
}
\arguments{
\item{data}{\code{Zoo}/\code{xts} object.}

\item{variable_of_interest}{Integer value indicating the index of the single target variable}

\item{fcast_horizon}{Integer value indicating the target horizon}

\item{delay}{Integer vector of predictor delays.}

\item{frequency}{Integer vector of frequencies of the variables in the data set (currently supported: \code{12} for monthly and \code{4} for quarterly data)}

\item{no_of_factors}{Integer number of factors.}

\item{seed}{32-bit positive integer for drawing the random hyper-parameter candidates}

\item{min_ridge_penalty}{Lower bound for the sampled ridge penalty coefficient}

\item{max_ridge_penalty}{Upper bound for the sampled ridge penalty coefficient}

\item{cv_repititions}{Integer number of predictions ought to be computed for each candidate set}

\item{cv_size}{Integer number of candidate sets}

\item{lasso_penalty_type}{Character indicating the lasso penalty type. \code{"selected"} uses the number of non-zero elements of the loading matrix. \code{"penalty"} uses the lasso size constraint directly. \code{"steps"} uses the number of steps.}

\item{min_max_penalty}{Vector of size tow, where the first element indicates the lower and the second element indicates the upper bound of the lasso penalty equivalent. If lasso_penalty_type is set to \code{"selected"} or \code{"steps"}, both elements must be strictly positive integers.}

\item{max_factor_lag_order}{Integer max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{lag_estim_criterion}{Information criterion used for the estimation of the factor VAR order (\code{"BIC"}, \code{"AIC"}, \code{"HIC"}).}

\item{ridge_penalty}{Ridge penalty.}

\item{lasso_penalty}{Numeric vector, lasso penalties for each factor (set to \code{NaN} if not intended as stopping criterion).}

\item{max_iterations}{Integer maximum number of iterations.}

\item{max_no_steps}{Integer number of max_no_steps (used for LARS-EN as an alternative).}

\item{comp_null}{Computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{conv_crit}{Conversion criterion for the SPCA algorithm.}

\item{conv_threshold}{Conversion criterion for the coordinate descent algorithm.}

\item{parallel}{Logical, whether or not to run the cross-validation loop in parallel.}

\item{max_ar_lag_order}{Integer maximum number of lags of the target variable ought to be included in the nowcasting step}

\item{max_predictor_lag_order}{Integer maximum number of lags of the predictors ought to be included in the nowcasting step}
}
\description{
Cross-validate the hyper-parameter of the mixed-frequency SDFM
}
