% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factor_model}
\alias{factor_model}
\title{Example factor model dataset}
\format{
A \code{SimulData} containing the following elements:
\describe{
\item{data}{If \code{starting_date} is provided, a \code{zoo} object, else,
a (\code{no_of_vars} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding the
simulated data.}
\item{factors}{If \code{starting_date} is provided, a \code{zoo} object, else a
(\code{no_of_factors} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding  the
simulated latent factors.}
\item{trans_var_coeff}{Numeric (\code{no_of_factors} \eqn{\times}{x}
(\code{no_of_factors} * \code{factor_lag_order})) factor VAR coefficient matrix.}
\item{loading_matrix}{Numeric factor loading matrix.}
\item{meas_error}{If \code{starting_date} is provided, a \code{zoo} object, else
a (\code{no_of_vars} \eqn{\times}{x} \code{no_of_obs}) numeric matrix holding the
fundamental measurement errors.}
\item{meas_error_var_cov}{Numeric measurement error
variance-covariance matrix.}
\item{trans_error_var_cov}{Numeric transition error
variance-covariance matrix.}
\item{frequency}{Integer vector of variable frequencies.}
\item{delay}{Integer vector of variable delays, measured as the number of
months since the latest available observation.}
}
}
\source{
Generated via \code{simFM()}. For details see \code{factor_model$call}.
}
\usage{
factor_model
}
\description{
This is a simulated factor model dataset for demonstration and testing
with \code{TwoStepSDFM}.
}
\keyword{datasets}
