% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateFactors.R
\name{noOfFactors}
\alias{noOfFactors}
\title{Estimate the number of factors using the Onatski (2010) procedure.}
\usage{
noOfFactors(
  data,
  min_no_factors = 1,
  max_no_factors = 15,
  confidence_threshold = 0.05
)
}
\arguments{
\item{data}{Numeric (no_of_variables x no_of_observations) matrix of data or zoo/xts object.}

\item{min_no_factors}{Integer minimum number of factors to be tested.}

\item{max_no_factors}{Integer maximum number of factors to be tested (should be at most min_no_factors + 17).}

\item{confidence_threshold}{Numeric threshold value to stop the testing procedure.}
}
\value{
Returns a list with the number of factors and some additional information.
}
\description{
Estimate the number of factors using the Onatski (2010) procedure.
}
