% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcast/forecast mixed-frequency data using an SDFM}
\usage{
nowcast(
  data,
  variables_of_interest,
  max_fcast_horizon,
  delay,
  selected,
  frequency,
  no_of_factors,
  sparse = TRUE,
  max_factor_lag_order = 10,
  decorr_errors = TRUE,
  lag_estim_criterion = "BIC",
  ridge_penalty = 1e-06,
  lasso_penalty = NULL,
  max_iterations = 1000,
  max_no_steps = NULL,
  comp_null = 1e-15,
  check_rank = FALSE,
  conv_crit = 1e-04,
  conv_threshold = 1e-04,
  log = FALSE,
  parallel = FALSE,
  max_ar_lag_order = 5,
  max_predictor_lag_order = 5
)
}
\arguments{
\item{data}{Zoo/xts object.}

\item{variables_of_interest}{Integer vector indicating the index of all target variables}

\item{max_fcast_horizon}{Maximum forecasting horizon}

\item{delay}{Integer vector of predictor delays.}

\item{selected}{Integer vector of the number of selected variables for each factor.}

\item{frequency}{Integer vector of frequencies of the variables in the data set (currently supported: "12" for monthly and "4" for quarterly data)}

\item{no_of_factors}{Integer number of factors.}

\item{sparse}{Logical, indicating whether a sparse model should be used for nowcasting}

\item{max_factor_lag_order}{Integer max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{lag_estim_criterion}{Information criterion used for the estimation of the factor VAR order ("BIC", "AIC", "HIC").}

\item{ridge_penalty}{Ridge penalty.}

\item{lasso_penalty}{Numeric vector, lasso penalties for each factor (set to NaN if not intended as stopping criterion).}

\item{max_iterations}{Integer maximum number of iterations.}

\item{max_no_steps}{Integer number of max_no_steps (used for LARS-EN as an alternative).}

\item{comp_null}{Computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{conv_crit}{Conversion criterion for the SPCA algorithm.}

\item{conv_threshold}{Conversion criterion for the coordinate descent algorithm.}

\item{log}{Logical, whether or not output should be printed along the algorithm.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel matrix operations.}

\item{max_ar_lag_order}{Integer maximum number of lags of the target variable ought to be included in the nowcasting step}

\item{max_predictor_lag_order}{Integer maximum number of lags of the predictors ought to be included in the nowcasting step}
}
\value{
The \code{nowcast} function returns an object which, contains the following elements:
\describe{
\item{\code{loading_matrix_estimate}}{A matrix of estimated loadings for each factor on each observed variable.}
\item{\code{smoothed_factors}}{The smoothed factor estimates.}
\item{\code{smoothed_state_variance}}{The estimated variance-covariance matrix of the residuals/errors.}
\item{\code{factor_var_lag_order}}{The order of the VAR(max_factor_lag_order) model used in the factor model process.}
\item{\code{companion_form_smoothed_factors}}{The smoothed factor estimates in companion form.}
\item{\code{error_var_cov_cholesky_factor}}{A matrix representing any transformations applied to the factors for error decorrelation, if \code{decorr_errors} is TRUE.}
}
}
\description{
Nowcast/forecast mixed-frequency data using an SDFM
}
