% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{plot.SDFMcrossVal}
\alias{plot.SDFMcrossVal}
\title{Generic plotting function for SDFMnowcast S3 objects}
\usage{
\method{plot}{SDFMcrossVal}(x, ...)
}
\arguments{
\item{x}{\code{SDFMcrossVal} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic plotting function for SDFMnowcast S3 objects
}
