% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{plot.SDFMFit}
\alias{plot.SDFMFit}
\title{Generic plotting function for SDFMFit S3 objects}
\usage{
\method{plot}{SDFMFit}(x, ...)
}
\arguments{
\item{x}{\code{SDFMFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic plotting function for SDFMFit S3 objects
}
