% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nowcast.R
\name{plot.SDFMnowcast}
\alias{plot.SDFMnowcast}
\title{Generic plotting function for SDFMnowcast S3 objects}
\usage{
\method{plot}{SDFMnowcast}(x, ...)
}
\arguments{
\item{x}{\code{SDFMnowcast} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic plotting function for SDFMnowcast S3 objects
}
