% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{print.SDFMFit}
\alias{print.SDFMFit}
\title{Generic printing function for SDFMFit S3 objects}
\usage{
\method{print}{SDFMFit}(x, ...)
}
\arguments{
\item{x}{\code{SDFMFit} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic printing function for SDFMFit S3 objects
}
