% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nowcast.R
\name{print.SDFMnowcast}
\alias{print.SDFMnowcast}
\title{Generic print function for SDFMnowcast S3 objects}
\usage{
\method{print}{SDFMnowcast}(x, ...)
}
\arguments{
\item{x}{\code{SDFMnowcast} object.}

\item{...}{Additional parameters for the plotting functions.}
}
\description{
Generic print function for SDFMnowcast S3 objects
}
