% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{SimFM}
\alias{SimFM}
\title{Draw data from a dynamic factor model}
\usage{
SimFM(
  T,
  N,
  R,
  Lambda,
  mu_xi,
  Sigma_xi,
  Sigma_epsilon,
  Phi,
  P,
  quarterfy,
  m = 0,
  corr = FALSE,
  beta_param = Inf,
  seed = 20022024,
  burn_in = 1000,
  rescale = TRUE,
  check_staionarity = FALSE,
  stationarity_check_threshold = 1e-05
)
}
\arguments{
\item{T}{Number of observations.}

\item{N}{Number of Variables.}

\item{R}{Number of factors.}

\item{Lambda}{Loadings matrix.}

\item{mu_xi}{Mean of the measurement errors.}

\item{Sigma_xi}{Variance-covariance matrix of the measurement errors.}

\item{Sigma_epsilon}{Variance-covariance matrix of the transition errors.}

\item{Phi}{Either a list of length P or a matrix of dimensions Rx(R * P) or (R * P)xR holding the VAR coefficients of the factor VAR process.}

\item{P}{Order of the VAR process in the state equation.}

\item{quarterfy}{Whether or not some of the data should be aggregated to quarterly representations.}

\item{m}{Ratio of variables ought to be quarterfied.}

\item{corr}{Whether or not the measurement error should be randomly correlated inside the function using a random correlation matrix with off-diagonal elements governed by a beta-distribution.}

\item{beta_param}{Parameter of the beta-distribution governing the off-diagonal elements of the variance-covariance matrix of the measurement error.}

\item{seed}{Seed for all random processes inside the function.}

\item{burn_in}{Burn-in period of the simulated data ought to be discarded at the beginning of the data.}

\item{rescale}{Whether or not the variance of the measurement error should be rescaled by the common component to equalise the signal-to-noise ratio.}

\item{stationarity_check_threshold}{Threshold of the stationarity check for when to deem an eigenvalue negative beyond the numerical error.}

\item{check_stationarity}{Whether or not the stationarity properties of the factor VAR process should be checked.}
}
\value{
Returns an \code{FM} object containing the DFM parameters expressed in the factor VAR companion form.
The object includes:
\describe{
\item{\code{F}}{A \eqn{RP \times T} matrix of simulated factors.}
\item{\code{Phi}}{The \eqn{R \times RP} factor VAR coefficients matrix.}
\item{\code{Lambda}}{The \eqn{N \times RP} loadings matrix.}
\item{\code{Sigma_xi}}{The \eqn{N \times N} variance-covariance matrix of the measurement errors.}
\item{\code{Sigma_epsilon}}{The \eqn{R \times R} variance-covariance matrix of the transition errors.}
\item{\code{Xi}}{A \eqn{N \times T} matrix representing the measurement errors.}
\item{\code{X}}{The \eqn{N \times T} matrix of simulated observed variables.}
\item{\code{frequency}}{The \eqn{N \times 1} vector of frequency of the data.}
}
}
\description{
The function simulates an (approximate) Dynamic Factor Model
}
