% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{simFM}
\alias{simFM}
\title{Draw data from a dynamic factor model

The function simulates an (approximate) Dynamic Factor Model}
\usage{
simFM(
  no_of_observations,
  no_of_variables,
  no_of_factors,
  loading_matrix,
  meas_error_mean,
  meas_error_var_cov,
  trans_error_var_cov,
  trans_var_coeff,
  factor_lag_order,
  quarterfy,
  quarterly_variable_ratio = 0,
  corr = FALSE,
  beta_param = Inf,
  seed = 20022024,
  burn_in = 1000,
  rescale = TRUE,
  check_stationarity = FALSE,
  stationarity_check_threshold = 1e-05,
  parallel = FALSE
)
}
\arguments{
\item{no_of_observations}{Number of observations.}

\item{no_of_variables}{Number of Variables.}

\item{no_of_factors}{Number of factors.}

\item{loading_matrix}{Loadings matrix.}

\item{meas_error_mean}{Mean of the measurement errors.}

\item{meas_error_var_cov}{Variance-covariance matrix of the measurement errors.}

\item{trans_error_var_cov}{Variance-covariance matrix of the transition errors.}

\item{trans_var_coeff}{Either a list of length max_factor_lag_order or a matrix of dimensions Rx(no_of_factors * max_factor_lag_order) or (no_of_factors * max_factor_lag_order)xR holding the VAR coefficients of the factor VAR process.}

\item{factor_lag_order}{Order of the VAR process in the state equation.}

\item{quarterfy}{Whether or not some of the data should be aggregated to quarterly representations.}

\item{quarterly_variable_ratio}{Ratio of variables ought to be quarterfied.}

\item{corr}{Whether or not the measurement error should be randomly correlated inside the function using a random correlation matrix with off-diagonal elements governed by a beta-distribution.}

\item{beta_param}{Parameter of the beta-distribution governing the off-diagonal elements of the variance-covariance matrix of the measurement error.}

\item{seed}{Seed for all random processes inside the function.}

\item{burn_in}{Burn-in period of the simulated data ought to be discarded at the beginning of the data.}

\item{rescale}{Whether or not the variance of the measurement error should be rescaled by the common component to equalise the signal-to-noise ratio.}

\item{check_stationarity}{Whether or not the stationarity properties of the factor VAR process should be checked.}

\item{stationarity_check_threshold}{Threshold of the stationarity check for when to deem an eigenvalue negative beyond the numerical error.}

\item{parallel}{Make use of Eigen internal parallel matrix operations}
}
\value{
Returns an \code{simFM} object containing the DFM parameters expressed in the factor VAR companion form.
}
\description{
Draw data from a dynamic factor model

The function simulates an (approximate) Dynamic Factor Model
}
