% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFM.r
\name{simFM}
\alias{simFM}
\title{Draw data from a dynamic factor model.}
\usage{
simFM(
  no_of_observations,
  no_of_variables,
  no_of_factors,
  loading_matrix,
  meas_error_mean,
  meas_error_var_cov,
  trans_error_var_cov,
  trans_var_coeff,
  factor_lag_order,
  delay = NULL,
  quarterfy = FALSE,
  quarterly_variable_ratio = 0,
  corr = FALSE,
  beta_param = Inf,
  seed = 20022024,
  burn_in = 1000,
  rescale = TRUE,
  starting_date = NULL,
  check_stationarity = FALSE,
  stationarity_check_threshold = 1e-05,
  parallel = FALSE
)
}
\arguments{
\item{no_of_observations}{Integer number of observations.}

\item{no_of_variables}{Integer number of Variables.}

\item{no_of_factors}{Integer number of factors.}

\item{loading_matrix}{Numeric (no_of_variables x no_of_factors) loading matrix.}

\item{meas_error_mean}{Numeric vector/matrix of the means of the measurement errors.}

\item{meas_error_var_cov}{Numeric (no_of_factors x no_of_factors) variance-covariance matrix of the measurement errors.}

\item{trans_error_var_cov}{Numeric (no_of_variables x no_of_variables) variance-covariance matrix of the transition errors.}

\item{trans_var_coeff}{Either a list of length max_factor_lag_order with each entry a numeric (no_of_factors x no_of_factors) VAR coefficient matrix or a matrix of dimensions (no_of_factors x(no_of_factors * max_factor_lag_order)) holding the VAR coefficients of the factor VAR process in each (no_of_factors x no_of_factors) block.}

\item{factor_lag_order}{Integer order of the VAR process in the state equation.}

\item{delay}{Integer vector of delays imposed onto the end of the data in months (ragged edges)}

\item{quarterfy}{Logical, whether or not some of the data should be aggregated to quarterly representations.}

\item{quarterly_variable_ratio}{Ratio of variables ought to be quarterfied.}

\item{corr}{Logical, whether or not the measurement error should be randomly correlated inside the function using a random correlation matrix with off-diagonal elements governed by a beta-distribution.}

\item{beta_param}{Parameter of the beta-distribution governing the off-diagonal elements of the variance-covariance matrix of the measurement error.}

\item{seed}{32-bit unsigned integer seed for all random processes inside the function.}

\item{burn_in}{Integer burn-in period of the simulated data ought to be discarded at the beginning of the data.}

\item{rescale}{Logical, whether or not the variance of the measurement error should be rescaled by the common component to equalise the signal-to-noise ratio.}

\item{starting_date}{A date type object indicating the start of the dataset. If NULL (default), the function returns matrices with observations along the second dimension (i.e., time in columns). If specified, the function treats the data as a time series, aligning it accordingly.}

\item{check_stationarity}{Logical, whether or not the stationarity properties of the factor VAR process should be checked.}

\item{stationarity_check_threshold}{Threshold of the stationarity check for when to deem an eigenvalue negative beyond the numerical error.}

\item{parallel}{Logical, make use of Eigen internal parallel matrix operations}
}
\value{
Returns an \code{simFM} object containing the DFM parameters expressed in the factor VAR companion form.
The \code{simFM} function returns a \code{simFM} object which, contains the following elements:
\describe{
\item{\code{`data`}}{If \code{starting_date} is provided, a \code{zoo} matrix time series. Else, \code{no_of_variables} x \code{no_of_observations} simulated data matrix.}
\item{\code{`factors`}}{If \code{starting_date} is provided, a \code{zoo} matrix time series. Else, \code{no_of_factors} x \code{no_of_observations} simulated factor matrix.}
\item{\code{`factor_var_coeff_matrix`}}{\code{no_of_factors} x (\code{no_of_factors} * \code{factor_lag_order}) factor VAR coefficient matrix.}
\item{\code{`loading_matrix`}}{Factor loading matrix.}
\item{\code{`measurement_error`}}{If \code{starting_date} is provided, a \code{zoo} matrix time series. Else,\code{no_of_variables} x \code{no_of_observations} simulated measurement error matrix.}
\item{\code{`measurement_error_var_cov`}}{Measurement error variance covariance matrix.}
\item{\code{`transition_error_var_cov`}}{Factor VAR process error (transition error) variance covariance matrix.}
\item{\code{`frequency`}}{Vector of variable frequencies}
\item{\code{`delay`}}{Vector of variable delays}
}
}
\description{
Draw data from a dynamic factor model.
}
