% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepDenseDFM.R
\name{twoStepDenseDFM}
\alias{twoStepDenseDFM}
\title{Estimate an DFM via PCA and Kalman Filter and Smoother according to Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic factor models based on Kalman filtering. Journal of Econometrics, 164(1), 188-205.}
\usage{
twoStepDenseDFM(
  data,
  delay,
  no_of_factors,
  max_factor_lag_order = 10,
  decorr_errors = TRUE,
  lag_estim_criterion = "BIC",
  comp_null = 1e-15,
  check_rank = FALSE,
  log = FALSE,
  parallel = FALSE,
  fcast_horizon = 0
)
}
\arguments{
\item{data}{Numeric (no_of_variables x no_of_observations) matrix of data or zoo/xts object.}

\item{delay}{Integer vector of variable delays.}

\item{no_of_factors}{Integer number of factors.}

\item{max_factor_lag_order}{Integer max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{lag_estim_criterion}{Information criterion used for the estimation of the factor VAR order ("BIC", "AIC", "HIC").}

\item{comp_null}{Computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{log}{Logical, whether or not output should be printed along the algorithm.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel matrix operations.}

\item{fcast_horizon}{Integer forecasting horizon for factor forecasts}
}
\value{
The \code{twoStepDenseDFM} function returns an object \code{fit} which, contains the following elements:
\describe{
\item{\code{loading_matrix_estimate}}{A matrix of estimated loadings for each factor on each observed variable.}
\item{\code{smoothed_factors}}{The smoothed factor estimates.}
\item{\code{smoothed_state_variance}}{The estimated variance-covariance matrix of the residuals/errors.}
\item{\code{factor_var_lag_order}}{The order of the VAR(max_factor_lag_order) model used in the factor model process.}
\item{\code{error_var_cov_cholesky_factor}}{A matrix representing any transformations applied to the factors for error decorrelation, if \code{decorr_errors} is TRUE.}
}
}
\description{
Estimate an DFM via PCA and Kalman Filter and Smoother according to Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic factor models based on Kalman filtering. Journal of Econometrics, 164(1), 188-205.
}
