% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepDenseDFM.R
\name{twoStepDenseDFM}
\alias{twoStepDenseDFM}
\title{Estimate a DFM via PCA and Kalman Filter and Smoother.}
\usage{
twoStepDenseDFM(
  data,
  delay,
  no_of_factors,
  max_factor_lag_order = 10,
  decorr_errors = TRUE,
  lag_estim_criterion = "BIC",
  comp_null = 1e-15,
  check_rank = FALSE,
  log = FALSE,
  parallel = FALSE,
  fcast_horizon = 0
)
}
\arguments{
\item{data}{Numeric (no_of_variables x no_of_observations) matrix of data or zoo/xts object.}

\item{delay}{Integer vector of variable delays.}

\item{no_of_factors}{Integer number of factors.}

\item{max_factor_lag_order}{Integer max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{lag_estim_criterion}{Information criterion used for the estimation of the factor VAR order ("BIC", "AIC", "HIC").}

\item{comp_null}{Computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{log}{Logical, whether or not output should be printed along the algorithm.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel matrix operations.}

\item{fcast_horizon}{Integer forecasting horizon for factor forecasts}
}
\value{
An object of class \code{DFMFit} with components:
\describe{
\item{data}{Original data matrix.}
\item{loading_matrix_estimate}{Matrix of estimated factor loadings.}
\item{smoothed_factors}{Matrix or \code{zoo} object containing the smoothed latent factors.}
\item{smoothed_state_variance}{Matrix with the state covariance matrices corresponding to the smoothed state vector (stacked in column-major order).}
\item{factor_var_lag_order}{Integer, selected lag order of the factor VAR model.}
\item{error_var_cov_cholesky_factor}{Lower-triangular Cholesky factor of the estimated measurement error variance–covariance matrix.}
}
}
\description{
Estimate a dense (approximate) dynamic factor model via principal components
analysis and a state-space representation with Kalman filter and smoother,
following Giannone, Reichlin and Small (2008) \url{doi:10.1016/j.jmoneco.2008.05.010}
and Doz, Giannone and Reichlin (2011) \url{doi:10.1016/j.jeconom.2011.02.012}.
}
\examples{
\dontrun{
set.seed(02102025)
no_of_observations <- 100
no_of_variables <- 20
no_of_factors <- 3
trans_error_var_cov <- diag(1, no_of_factors)
loading_matrix <- matrix(round(rnorm(no_of_variables * no_of_factors)), no_of_variables, no_of_factors)
meas_error_mean <- rep(0, no_of_variables)
meas_error_var_cov <- diag(1, no_of_variables)
trans_var_coeff <- cbind(diag(0.5, no_of_factors), -diag(0.25, no_of_factors))
factor_lag_order <- 2
FM <- simFM(no_of_observations = no_of_observations, no_of_variables = no_of_variables, 
            no_of_factors = no_of_factors, loading_matrix = loading_matrix, 
            meas_error_mean = meas_error_mean, meas_error_var_cov = meas_error_var_cov,
            trans_error_var_cov = trans_error_var_cov, trans_var_coeff = trans_var_coeff, 
            factor_lag_order = factor_lag_order)

# Estimate a dense 2-factor model
fit <- twoStepDenseDFM(
  data = FM$data,
  delay = FM$delay,
  no_of_factors = no_of_factors
)

print(fit)
plots <- plot(fit)
}

}
