% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{twoStepSDFM}
\alias{twoStepSDFM}
\title{Estimate an SDFM via SPCA and Kalman Filter and Smoother}
\usage{
twoStepSDFM(
  data,
  delay,
  selected,
  no_of_factors,
  max_factor_lag_order = 10,
  decorr_errors = TRUE,
  lag_estim_criterion = "BIC",
  ridge_penalty = 1e-06,
  lasso_penalty = NULL,
  max_iterations = 1000,
  max_no_steps = NULL,
  comp_null = 1e-15,
  check_rank = FALSE,
  conv_crit = 1e-04,
  conv_threshold = 1e-04,
  log = FALSE,
  parallel = FALSE,
  fcast_horizon = 0
)
}
\arguments{
\item{data}{Numeric (no_of_variables x no_of_observations) matrix of data or zoo/xts object.}

\item{delay}{Integer vector of variable delays.}

\item{selected}{Integer vector of the number of selected variables for each factor.}

\item{no_of_factors}{Integer number of factors.}

\item{max_factor_lag_order}{Integer max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{lag_estim_criterion}{Information criterion used for the estimation of the factor VAR order ("BIC", "AIC", "HIC").}

\item{ridge_penalty}{Ridge penalty.}

\item{lasso_penalty}{Numeric vector, lasso penalties for each factor (set to NaN if not intended as stopping criterion).}

\item{max_iterations}{Integer maximum number of iterations.}

\item{max_no_steps}{Integer number of max_no_steps (used for LARS-EN as an alternative).}

\item{comp_null}{Computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{conv_crit}{Conversion criterion for the SPCA algorithm.}

\item{conv_threshold}{Conversion criterion for the coordinate descent algorithm.}

\item{log}{Logical, whether or not output should be printed along the algorithm.}

\item{parallel}{Logical, whether or not to use Eigen's internal parallel matrix operations.}

\item{fcast_horizon}{Integer forecasting horizon for factor forecasts}
}
\value{
An object of class \code{SDFMFit} with components:
\describe{
\item{data}{Original data matrix.}
\item{loading_matrix_estimate}{Matrix of estimated factor loadings.}
\item{smoothed_factors}{Matrix or \code{zoo} object containing the smoothed latent factors.}
\item{smoothed_state_variance}{Matrix with the state covariance matrices corresponding to the smoothed state vector (stacked in column-major order).}
\item{factor_var_lag_order}{Integer, selected lag order of the factor VAR model.}
\item{error_var_cov_cholesky_factor}{Lower-triangular Cholesky factor of the estimated measurement error variance–covariance matrix.}
}
}
\description{
Estimate a sparse mixed-frequency Gaussian dynamic factor model using
sparse principal components analysis and a state-space representation
with Kalman filter and smoother. For more details see Franjic and Schweikert (2024) \url{doi:10.2139/ssrn.4733872}.
}
\examples{
\dontrun{
set.seed(02102025)
no_of_observations <- 100
no_of_variables <- 20
no_of_factors <- 3
trans_error_var_cov <- diag(1, no_of_factors)
loading_matrix <- matrix(round(rnorm(no_of_variables * no_of_factors)), no_of_variables, no_of_factors)
meas_error_mean <- rep(0, no_of_variables)
meas_error_var_cov <- diag(1, no_of_variables)
trans_var_coeff <- cbind(diag(0.5, no_of_factors), -diag(0.25, no_of_factors))
factor_lag_order <- 2
FM <- simFM(no_of_observations = no_of_observations, no_of_variables = no_of_variables, 
            no_of_factors = no_of_factors, loading_matrix = loading_matrix, 
            meas_error_mean = meas_error_mean, meas_error_var_cov = meas_error_var_cov,
            trans_error_var_cov = trans_error_var_cov, trans_var_coeff = trans_var_coeff, 
            factor_lag_order = factor_lag_order)

# Estimate a 2-factor model
fit <- twoStepSDFM(
  data = FM$data,
  delay = FM$delay,
  selected = rep(floor(0.2 * no_of_variables), no_of_factors),
  no_of_factors = no_of_factors
)

# Inspect the results
print(fit)
plots <- plot(fit)
}

}
