% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepSDFM.R
\name{TwoStepSDFM}
\alias{TwoStepSDFM}
\title{Run Sparse DFM KFS}
\usage{
TwoStepSDFM(
  X,
  delay,
  selected,
  R,
  P = 10,
  decorr_errors = TRUE,
  crit = "BIC",
  l2 = 1e-06,
  l1 = NaN,
  max_iterations = 1000,
  steps = NaN,
  comp_null = 1e-15,
  check_rank = FALSE,
  conv_crit = 1e-04,
  conv_threshold = 1e-04,
  log = FALSE
)
}
\arguments{
\item{X}{Numeric "N x T" matrix of data.}

\item{delay}{Integer vector of predictor delays.}

\item{selected}{Integer vector of the number of selected variables for each factor.}

\item{R}{Integer, number of factors.}

\item{P}{Integer, max P of the VAR(P) process of the factors.}

\item{decorr_errors}{Logical, whether or not the errors should be decorrelated.}

\item{crit}{Character string, information criterion used for the estimation.}

\item{l2}{Numeric, ridge penalty.}

\item{l1}{Numeric vector, lasso penalties for each factor (set to NaN if not intended as stopping criterion).}

\item{max_iterations}{Integer, maximum number of iterations.}

\item{steps}{Integer, number of steps (used for LARS-EN as an alternative).}

\item{comp_null}{Numeric, computational zero.}

\item{check_rank}{Logical, whether or not the rank of the variance-covariance matrix should be checked.}

\item{conv_crit}{Numeric, conversion criterion for the SPCA algorithm.}

\item{conv_threshold}{Numeric, conversion criterion for the coordinate descent algorithm.}

\item{log}{Logical, whether or not output should be printed along the algorithm.}

\item{method}{Character string, method used for estimation ("BIC", "AIC", "HIC").}
}
\value{
The \code{TwoStepSDFM} function returns an object \code{fit} which, contains the following elements:
\describe{
\item{\code{Lambda_hat}}{A matrix of estimated loadings for each factor on each observed variable.}
\item{\code{F}}{The factor estimates.}
\item{\code{Pt}}{The estimated precision matrix of the factors.}
\item{\code{Wt}}{The estimated variance-covariance matrix of the residuals/errors.}
\item{\code{C}}{A matrix representing any transformations applied to the factors for error decorrelation, if \code{decorr_errors} is TRUE.}
\item{\code{P}}{The order of the VAR(P) model used in the factor model process.}
}
}
\description{
This function wraps the C++ implementation of the Sparse DFM KFS to
provide default parameters and potentially pre-process inputs or
post-process outputs.
}
\examples{
\dontrun{
result <- TwoStepSDFM(X = matrix_data, delay = delays, selected = selections)
}
}
